+++ AMaze/Entities/Key.cs
@@ -1,21 +1,19 @@
 ﻿namespace AMaze.Entities;
 
-internal class Key : IEnity
+internal class Key : IEntity
 {
 	public Geometry.IGeom Geom { get; }
-	public Wall Door { get; }
 
-	private bool pickedUp;
+	public event Action<Key>? PickupCallback;
 
-	public Key(double x, double y, Wall door)
+	public Key(double x, double y)
 	{
 		Geom = new Geometry.Face(x, y, 0.25);
-		Door = door;
 	}
 
 	public bool Intersect(Geometry.Seg ray, out ((double, double), ScanIntersectionExtra) intersection)
 	{
-		if (!pickedUp && Geom.Intersect(ray, out var point))
+		if (Geom.Intersect(ray, out var point))
 		{
 			var extra = new ScanIntersectionExtra { top = -0.5, bottom = -1.0, altPalette = true };
 			intersection = (point, extra);
@@ -27,12 +25,8 @@ internal class Key : IEnity
 
 	public bool DoesCollide(Geometry.Rect rect)
 	{
-		if (!pickedUp && Geom.DoesIntersect(rect))
-		{
-			pickedUp = true;
-			Door.IsGhost = true;
-			Door.IsVisible = false;
-		}
+		if (Geom.DoesIntersect(rect))
+			PickupCallback?.Invoke(this);
 		return false;
 	}
 }