+++ AMaze/ScanIntersectionExtra.cs
@@ -3,5 +3,5 @@
 internal struct ScanIntersectionExtra
 {
 	public double top, bottom;
-	public bool opaque, altPalette;
+	public bool opaque, altPalette, vantablack;
 }